Make Slow Requests an extension point.
